third_party/WebKit/LayoutTests/media/video-seek-no-src-expected.txt

Test that seeking video with no 'src' attribute sets default playback start position to that time.

EXPECTED (video.networkState == '0') OK
EXPECTED (video.readyState == '0') OK
EXPECTED (video.currentTime == '1.1') OK
END OF TEST
